requirements/base.txt
Django==5.1.7
django-environ==0.11.2
httpx==0.28.1
PyMySQL==1.1.1
